0.35 - using TC4 for step count, which allows faster counting. Added debug command.

+++ firmware/stepper_nano_zero/steppin.h
@@ -4,7 +4,7 @@
 
 void stepPinSetup(void); //setup step pin
 
-int32_t getSteps(void); //returns the number of steps changed since last call
+int64_t getSteps(void); //returns the number of steps changed since last call
 
 
 #endif // __STEPPIN_H___
